Update calculations/variance.go
Added comments

+++ calculations/variance.go
@@ -1,6 +1,6 @@
 package calculations
 
-func Variance(numbers []int) float64 {
+func Variance(numbers []int) int {
 	if len(numbers) == 0 {
 		return 0 // Avoid division by zero
 	}
@@ -8,10 +8,10 @@ func Variance(numbers []int) float64 {
 	n := len(numbers)
 	ave := Average(numbers) // Call average function
 
-	sum := 0.0
+	sum := 0
 	for i := 0; i < n; i++ {
-		sum += float64((numbers[i] - ave) * (numbers[i] - ave)) // Squared difference
+		sum += (numbers[i] - ave) * (numbers[i] - ave) // Squared difference
 	}
 
-	return sum / float64(n)
+	return sum / n
 }
